extracted make_ensemble and added it to test_preference_compairison.
